refactor(progress): widen current/total counters to u64 in progress contract (#4c 3.5)

+++ crates/quicksort-application/src/ports/outbound/progress_reporter.rs
@@ -11,10 +11,10 @@ use async_trait::async_trait;
 /// Progress information for a long-running operation.
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
 pub struct ProgressInfo {
-    /// Current step (0-based).
-    pub current: u32,
-    /// Total number of steps.
-    pub total: u32,
+    /// Current value of the progress meter (0-based).
+    pub current: u64,
+    /// Total value of the progress meter.
+    pub total: u64,
     /// Human-readable description of the current phase.
     pub phase: String,
     /// Optional detail message (e.g., filename being processed).
